[WIP] Restructure XC part to allow open shell
This is the first open shel version which compiles. The following has
been done:

1) XCPotential is now a base class on its own. It contains the
potential(s) of the XC functional and in the future will also include
its derivatives (hessian and further)
2) The open shell extension is achieved by including several potential
functions in XCPotential For instance, the Hessian will include one
function on the closed-shell case (second der. of the functionl wrt
total density), whereas in the open shell case il will include three
functions (alpha/alpha, alpha/beta and beta/beta derivatives
respectively)
3) XCOperator is no longer a pure virtual class but a real class, and
it has an XCPotential among its members. This allows to keep the
current code structure for SCF and Response. When applying the
XCOperator, the correct potential will have to be considered
(e.g. v_alpha or v_beta for alpha/beta orbitals in an SCF).