Fix memory deallocation problem in XCOperator
Now the XCOperator object contains a FunctionTreeVector for each of
the possible gradient-like input to xcfun: grad_a, grad_b, grad_t,
gamma. Therefore the xcInput array is now merely a book-keeping array
which is used to feed xcfun with functions in the correct order. As a
consequence, upon deallocation, the xcInput array shall NOT be
deallocated, but its elements set to zero. When the XCOperator is
either cleared or destrojed the the gradient functions are
deallocated.