Let mrchem run with inputs not in cwd
The user can now pass a path _relative_ to the current working directory
where the input file resides.
Note that it is assumed that the path is _relative_ to where the
script is invoked. If an absolute path is passed, it will still
be interpreted as relative causing trouble.
Moving to Python 3 would let us use pathlib which would make
my bad implementation much nicer, simpler, and general.