Add periodic Utils
Add on_boundary, check if nuclei is placed on the bondary

Add periodify_nuclei, if nuclei is placed on boundary, we can
use this to periodify it, from the POV of the code.

Add fix_boudary_charge, for nucleis placed on the boundary. This
can be used to ensure the total charge of the cell is correct.

Add calc_rc, finds the largest rc cutoff, depending on where the
nucleis are placed.